apvs-049 - Initial commit for Internal web.
Committing a copy of the gov.uk prototype kit for the internal website housed in a new directory.
Provided a gitignore file for this new module.
Set the dependencies in package.json to include redis.
Created a placeholder index.html page to serve as the landing page for internal users.
Altered the configuration such that the internal site launches on port 3001 (versus the default 3000 which is being used by the external site).