apvs-002 - Implemented form functionality for the escorts page.
Implemented get and post routes for the page.
Implemented the required form elements and tied them back to the persisted document for when the page is loaded.
Added ifs for each of the radio buttons to set them to the required value on page load.
Made the formatting of the HTML consistent.