Feature/apvs 002/complete external application form (#13)
* apvs-002 - Documentation improvements

Added example calls for the routes.
Reorganised the code to list the gets, and then the posts.

* apvs-002 - Fixed redirect to relationship.

When the user is redirected to the relationship page their claimant id is now included.

* apvs-002 - about-you page redirection now links correctly to relationship.

The insert method (when creating a new user) now retrieves the id of the new document and redirects the user along with the id to the relationship page.
Altered the syntax of the update methods the original implementation for update would overwrite the document stored in Mongo with whatever new object it was passed. Now only the values of the current document/form are updated.

* apvs-002 - Implemented form functionality for the relationship page.

Implemented get and post routes for the page.
Implemented the required form elements and tied them back to the persisted document for when the page is loaded.
Added ifs for each of the radio buttons to set them to the required value on page load.
Made the formatting of the HTML consistent.

* apvs-002 - Implemented form functionality for the escorts page.

Implemented get and post routes for the page.
Implemented the required form elements and tied them back to the persisted document for when the page is loaded.
Added ifs for each of the radio buttons to set them to the required value on page load.
Made the formatting of the HTML consistent.

* apvs-002 - Removed unnecessary console.log.

* apvs-002 - Implemented form functionality for the about-your-income page.

Implemented get and post routes for the page.
Implemented the required form elements and tied them back to the persisted document for when the page is loaded.
Added ifs for each of the radio buttons to set them to the required value on page load.
Made the formatting of the HTML consistent.