apvs-014 - Made approval, rejection, and escalation of eligibility claims functional. (#18)
Added status-partial.html that defines the view for displaying the three statuses and for updating the application status. Includes buttons for approving, rejecting, and escalating an eligibility claim.
Added routes for handling approve, reject, and escalate buttons on the status-partial.html.
The Data Table now displays the Application Status.
Cleaned up the index page, removed the placeholder breadcrumbs and made the table expand to fill more of the screen.
The initial pending status is now capitilised to match the case of the statuses set by internal web.