apvs-011 - Added implementation of the mock UI and API response for displaying NOMIS response.
Completed the relationship-partial.
Now has a button to update the relationship status on demand.
Added an API call to retrieve the stubbed NOMIS response, and update the relationship status in the database.
Fixed some bugs around the claimant-income-check page that meant it was saving the DWP response to the wrong part of the stored data. Paired with Paul on this.
Moved the partials to the views directory and renamed them to be consisting with the other status html file.