chore - Refactored the initial claimant save. (#31)
The save was composed of two database calls. One to save the claimant and a second to update the claimants status. This is now done in a single call.
The personal information captured on the about-you page is now grouped under the key 'personal' in the mongo document.
Updated all references to the personal information across both applications.